refactor: split horizontal scroll variants
